(PC-10647) fix bug andd show venue label on all pages when selected
